perf(editor): ship parse-reset-boundaries-v2 hybrid widener
Lands the openspec change `parse-reset-boundaries-v2` plus the
follow-on review fixes, instrumentation, refactor, and verify
demotion.

V2 hybrid widener
- ParsedBuffer.lazy_depth: per-row depth of open lazy-continuable
  constructs (List, BlockQuote, IndentedCode multi-chunk, HtmlBlock).
  Populated by a new lazy_delta accumulator parallel to the existing
  event walk; prefix-summed into u32 per row.
- byte_to_row_col_unclamped: dedicated past-EOF resolver for End
  events so the lazy_delta decrement lands in the off-by-one slot
  beyond `lines` instead of zeroing the last content row.
- reset_boundaries detection switched to `lazy_depth == 0 && Blank`
  so blanks inside lazy constructs are excluded.
- try_incremental_parse: hybrid widener tries
  expand_to_reset_boundary first (provably equivalent slice, no
  release verify); on cap-trip falls back to widen_to_safe (verify
  required in debug for correctness).

Structural guards (in execution order)
- Pre-edit kind check: FenceMarker / SetextUnderline / IndentedCode
  / HtmlBlock bail to full parse (v1, retained).
- looks_like_* flips for fence, setext, indented-code, html-block,
  list-marker, blockquote-marker (last two new in this change; the
  loose-list-merge case `["x","","* a","* b"]` + typing `* x` at
  row 0 was a real divergence the v1 guard set missed).
- lazy_depth[row-1..=row+1] > 0: edits adjacent to or inside a
  lazy construct could re-extend it across the splice.
- Blank<->non-blank flip with non-blank neighbour: invalidates the
  pre-edit boundary at the row in the post-edit world.
- looks_like_indented_code and looks_like_list_marker tightened
  to require non-whitespace content after the indent / marker
  (matching pulldown's actual recognition).

Bug fixes surfaced by extra-high-recall code review
- byte_to_row_col_unclamped previously mis-fired its past-EOF
  branch when a block ended at the START of a trailing empty row
  (`Ok(r)` exact-match case with a 0-length last row). Fix: split
  Ok/Err arms; only Err-with-past-content goes past-EOF.
  Regression: lazy_depth_blockquote_with_trailing_blank_drops_at_blank.
- debug_assert!(>=0) + .max(0) as u32 was a textbook masking pair
  (panic in debug, silent clamp in release). Dropped the clamp; the
  cast surfaces underflow loudly in release if the assert is ever
  violated.
- Vec<bool> indented_codeblock_stack -> Option<bool> + debug_assert
  on End-without-Start: CommonMark code blocks don't nest, so the
  type encodes the invariant and surfaces pulldown contract
  violations instead of silently swallowing them.
- byte_to_row_col_unclamped now called only inside the End arms
  that actually need it, not on every event in the walk.
- Dead `depth` prefix-sum + depth_delta vec + is_top_level_block_*
  helpers removed (their only consumer was the v1 boundary
  detector).
- MarkdownEditorView::new uses ParsedBuffer::placeholder(&[])
  instead of hand-rolling the literal so future field additions
  don't drift.
- old_blank derived from old_kind == LineConstructKind::Blank
  instead of recomputing `line.trim().is_empty()`.
- Debug-mode KIMUN_VIEW_VERIFY_INCREMENTAL verify extended to
  compare lazy_depth AND reset_boundaries against fresh parse.

Bounds guards on pulldown event arms
- Pre-existing OOB panic on Fenced/Indented/HtmlBlock arms when
  `sr == er` (single-row degenerate fence at EOF) or
  `er == lines.len()` (unclosed fence) — caused by raw
  `kinds[(sr+1)..er].fill(...)` slicing. Added bounded `(er+1)
  .min(kinds.len())` slicing to each arm. This panic blocked
  proptest soak before any v2 work could run.

Refactor: markdown.rs (2627 LOC) split into folder module
- markdown/mod.rs: Element/ElementKind/ImagePlaceholder/ParsedLine
  types, tag_to_kind, span_style, shared helpers.
- markdown/parsed_buffer.rs: ParsedBuffer + parse + splice +
  lazy_depth tracking + byte_to_row_col* + lazy-continuable
  helpers.
- markdown/spanner.rs: MarkdownSpanner + render/cursor/visible/
  rendered_col_to_logical + span_style.
- markdown/detect.rs: detect_wikilinks + detect_image_placeholders.

Widener metrics instrumentation
- widener_metrics.rs: atomic counters for every outcome category
  (incremental_reset, incremental_fallback, line_count_change,
  kind_guard, lazy_depth, blank_transition, cap_trip,
  verify_failed) plus raw entry/view-update counters.
- Each guard return calls METRICS.bail(reason); successes call
  METRICS.ok(path).
- KIMUN_DUMP_WIDENER_METRICS=1 prints session totals + derived
  metrics (successful_incremental_rate, fast_path_share,
  guard_sprawl_rate, verify_hit_rate) on clean app exit.
- IMPORTANT: dump routed via `crate::components::...` not
  `kimun_notes::...` because the lib + bin both declare
  `pub mod components` and end up with separate METRICS statics.
  Detailed inline comment in main.rs.

Verify demotion (v2 phase 7)
- Multi-session dogfooding across 8 buffer shapes and 1581
  attempts produced ZERO verify_failed events, including the
  widen_to_safe fallback path's only invocation. The post-slice
  verify loop is gated behind cfg(debug_assertions) +
  verify_incremental_enabled(). Debug builds (incl. proptest
  soak) still run it.

Sample fixtures + bench
- example/work/widener-stress/: 8 stress-test notes covering
  long-no-blank prose, code-heavy, heavy lists, blockquotes,
  indented-code multi-chunk, mixed realistic, heterogeneous
  lazy-dense, and short-simple shapes. Plus a README with
  collection workflow and per-shape signal interpretation.
- bench_full_view_update_heavy_lists_571_typing: 493 µs for a
  single-char edit inside a 571-row loose list (full-rebuild
  path). 3% of 60 Hz typing budget — not a real product issue
  at current sizes.

Proptest harness
- ASCII printable char strategy converted from prop_filter to a
  range strategy so 100k cases don't local-reject ceiling-trip.
- New incremental_matches_full_for_lazy_continuation_buffers
  proptest targets blockquote-containing-indented-code shapes
  from the v2 proposal.
- 100k cases pass with KIMUN_VIEW_VERIFY_INCREMENTAL=1.

Test results
- 1528 unit + 3 proptests (100k each) pass.
- Clippy clean.
- bench_full_view_update_5000_lines_incremental: ~240 µs, -36%
  from pre-v2 baseline.
