+++ example/work/projects/search-caching.md
