+++ example/work/widener-stress/heterogeneous_lazy_dense.md
@@ -1,0 +1,199 @@
+    indented code 1
+    second indented line
+
+- list item 2
+- list item 2 continued
+
+Plain paragraph 3 with text.
+
+> blockquote line 4
+
+    indented code 5
+    second indented line
+
+- list item 6
+- list item 6 continued
+
+Plain paragraph 7 with text.
+
+> blockquote line 8
+
+    indented code 9
+    second indented line
+
+- list item 10
+- list item 10 continued
+
+Plain paragraph 11 with text.
+
+> blockquote line 12
+
+    indented code 13
+    second indented line
+
+- list item 14
+- list item 14 continued
+
+Plain paragraph 15 with text.
+
+> blockquote line 16
+
+    indented code 17
+    second indented line
+
+- list item 18
+- list item 18 continued
+
+Plain paragraph 19 with text.
+
+> blockquote line 20
+
+    indented code 21
+    second indented line
+
+- list item 22
+- list item 22 continued
+
+Plain paragraph 23 with text.
+
+> blockquote line 24
+
+    indented code 25
+    second indented line
+
+- list item 26
+- list item 26 continued
+
+Plain paragraph 27 with text.
+
+> blockquote line 28
+
+    indented code 29
+    second indented line
+
+- list item 30
+- list item 30 continued
+
+Plain paragraph 31 with text.
+
+> blockquote line 32
+
+    indented code 33
+    second indented line
+
+- list item 34
+- list item 34 continued
+
+Plain paragraph 35 with text.
+
+> blockquote line 36
+
+    indented code 37
+    second indented line
+
+- list item 38
+- list item 38 continued
+
+Plain paragraph 39 with text.
+
+> blockquote line 40
+
+    indented code 41
+    second indented line
+
+- list item 42
+- list item 42 continued
+
+Plain paragraph 43 with text.
+
+> blockquote line 44
+
+    indented code 45
+    second indented line
+
+- list item 46
+- list item 46 continued
+
+Plain paragraph 47 with text.
+
+> blockquote line 48
+
+    indented code 49
+    second indented line
+
+- list item 50
+- list item 50 continued
+
+Plain paragraph 51 with text.
+
+> blockquote line 52
+
+    indented code 53
+    second indented line
+
+- list item 54
+- list item 54 continued
+
+Plain paragraph 55 with text.
+
+> blockquote line 56
+
+    indented code 57
+    second indented line
+
+- list item 58
+- list item 58 continued
+
+Plain paragraph 59 with text.
+
+> blockquote line 60
+
+    indented code 61
+    second indented line
+
+- list item 62
+- list item 62 continued
+
+Plain paragraph 63 with text.
+
+> blockquote line 64
+
+    indented code 65
+    second indented line
+
+- list item 66
+- list item 66 continued
+
+Plain paragraph 67 with text.
+
+> blockquote line 68
+
+    indented code 69
+    second indented line
+
+- list item 70
+- list item 70 continued
+
+Plain paragraph 71 with text.
+
+> blockquote line 72
+
+    indented code 73
+    second indented line
+
+- list item 74
+- list item 74 continued
+
+Plain paragraph 75 with text.
+
+> blockquote line 76
+
+    indented code 77
+    second indented line
+
+- list item 78
+- list item 78 continued
+
+Plain paragraph 79 with text.
+
+> blockquote line 80
