+++ example/work/widener-stress/short_simple.md
@@ -1,0 +1,9 @@
+# Quick note
+
+hola
+
+One line.
+
+Another line, separated.
+
+Done.
